Rename scm.libbase references to scm.base SO107
