+++ doc/source/examples/ZieglerNattaCatalyst/ZieglerNattaCatalyst.ipynb.rst
@@ -18,7 +18,7 @@ Initial imports
 
 .. code:: ipython3
 
-   from scm.libbase import ChemicalSystem
+   from scm.base import ChemicalSystem
    from scm.plams import view, Settings, AMSJob, Units
 
 Initial system
