+++ src/scm/plams/interfaces/adfsuite/inputparser.py
@@ -11,8 +11,8 @@ from scm.plams.interfaces.adfsuite.amsworker import AMSWorker, AMSWorkerError
 from scm.plams.interfaces.adfsuite.ams import AMSJob
 
 if TYPE_CHECKING:
-    from scm.libbase import InputParser as InputParserLibbase
-    from scm.libbase import InputFile as InputFileLibbase
+    from scm.base import InputParser as InputParserLibbase
+    from scm.base import InputFile as InputFileLibbase
 
 TSelf = TypeVar("TSelf", bound="InputParser")
 
@@ -23,15 +23,15 @@ class InputParser:
     """
     A utility class for converting text input into JSON dictionaries and plams.Settings.
 
-    This is a legacy implementation for environments without access to scm.libbase.
+    This is a legacy implementation for environments without access to scm.base.
     """
 
     # !!!!!!!  DEPRECATED  !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-    # This class has been deprecated. The equivalent in scm.libbase
-    # should be used instead where possible. The scm.libbase version does the input
+    # This class has been deprecated. The equivalent in scm.base
+    # should be used instead where possible. The scm.base version does the input
     # parsing via direct calls into libscm_base, instead of spawning an AMSWorker
     # and then pushing the input through the pipe. This implementation exists here
-    # only to remove the dependency on scm.libbase when running in python
+    # only to remove the dependency on scm.base when running in python
     # environments without access to the base library.
     # !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
@@ -79,28 +79,28 @@ class InputParserFacade:
     """
     A utility class for converting text input into JSON dictionaries and plams.Settings.
 
-    Uses the scm.libbase implementation of InputParser if available, and otherwise the legacy implementation
+    Uses the scm.base implementation of InputParser if available, and otherwise the legacy implementation
     of the parser which spawns an AMSWorker instance.
     """
 
     try:
-        from scm.libbase import InputParser as InputParserLibbase
+        from scm.base import InputParser as InputParserLibbase
 
         # Cache a single instance of the parser to avoid having to repeatedly reload input file definition JSON
         # But for this need to make access to the parser thread-safe
-        input_parser_scm_libbase = InputParserLibbase()
+        input_parser_scm_base = InputParserLibbase()
         input_parser_lock = threading.Lock()
-        _has_scm_libbase = True
+        _has_scm_base = True
     except ImportError:
-        _has_scm_libbase = False
+        _has_scm_base = False
 
     @property
     def parser(self) -> Union[InputParser, "InputParserLibbase"]:
         """
         Get instance of a parser used to convert text input.
         """
-        if self._has_scm_libbase:
-            return self.input_parser_scm_libbase
+        if self._has_scm_base:
+            return self.input_parser_scm_base
         else:
             return InputParser()
 
@@ -108,7 +108,7 @@ class InputParserFacade:
         """
         Run a string of text input through the input parser and produce a Python dictionary representing the JSONified input.
         """
-        if self._has_scm_libbase:
+        if self._has_scm_base:
             with self.input_parser_lock:
                 return self.parser.to_dict(program, text_input, string_leafs)
         else:
@@ -116,7 +116,7 @@ class InputParserFacade:
                 return parser.to_dict(program, text_input, string_leafs)
 
 
-@requires_optional_package("scm.libbase")
+@requires_optional_package("scm.base")
 def get_system_blocks_as_molecules_from_input(input_file: "InputFileLibbase") -> Dict[str, Molecule]:
     """
     Get a dictionary of mappings between the System blocks in the input to |Molecule| instances.
