+++ unit_tests/test_amsresults.py
@@ -247,7 +247,7 @@ class TestWaterOptimizationAMSResults:
 
         # When get molecule from given file section
         try:
-            from scm.libbase import ChemicalSystem
+            from scm.base import ChemicalSystem
 
             # Then molecule as expected when chemical system present
             molecule = water_opt_results.get_system(section, file)
@@ -288,7 +288,7 @@ class TestWaterOptimizationAMSResults:
 
         # When get input molecule
         try:
-            from scm.libbase import ChemicalSystem
+            from scm.base import ChemicalSystem
 
             # Then molecule as expected when chemical system present
             molecule = water_opt_results.get_input_system()
@@ -323,7 +323,7 @@ class TestWaterOptimizationAMSResults:
 
         # Then molecule as expected (post optimization)
         try:
-            from scm.libbase import ChemicalSystem
+            from scm.base import ChemicalSystem
 
             # Then molecule as expected when chemical system present
             molecule = water_opt_results.get_main_system()
